% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmDataFunction.R
\name{bbmData}
\alias{bbmData}
\title{bbmData Function}
\usage{
bbmData(modelDate = "12_27_2018", cookies = bbmCookies,
  refreshCookie = FALSE)
}
\arguments{
\item{modelDate}{"12_27_2018"}

\item{cookies}{bbmCookies}

\item{refreshCookie}{T/F(FALSE)}
}
\value{
bbm Data spreadsheet in NBA_Daily/bbMonster/Year/Date/ directory
}
\description{
bbmData Function
}
\examples{
bbmData(modelDate="!2_27_2018",cookies=bbmCookies,refreshCookie=FALSE)
}
