% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretTimesTestFunction.R
\name{caretTimesTest}
\alias{caretTimesTest}
\title{caret Times Test Function}
\usage{
caretTimesTest(model)
}
\arguments{
\item{model}{caretModel}
}
\value{
times
}
\description{
caret Times Test Function
}
\examples{
caretTimesTest(nbaZeroOwnrangerOOB)
}
