% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerColMeansFunction.R
\name{center_colmeans}
\alias{center_colmeans}
\title{colmeans centering}
\usage{
center_colmeans(x)
}
\arguments{
\item{x}{data}
}
\value{
mean-centered data for analysis
}
\description{
colmeans centering
}
\examples{
center_colmeans(batsOnly)
}
