% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearCombosFilterFunction.R
\name{combos.Filter}
\alias{combos.Filter}
\title{lcs removal}
\usage{
combos.Filter(x)
}
\arguments{
\item{x}{data}
}
\value{
dataframe
}
\description{
lcs removal
}
\examples{
combos.Filter(bats)
}
