% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateRangeFormFunction.R
\name{dateRangeForm}
\alias{dateRangeForm}
\title{labs dateRange Formatting Function}
\usage{
dateRangeForm(yyyymmdd = "2018/03/29", days = 33)
}
\arguments{
\item{yyyymmdd}{"2018/03/29"}

\item{days}{any# i.e:44}
}
\value{
list of dates for use
}
\description{
labs dateRange Formatting Function
}
\examples{
dateRnageForm("2018/03/29",days=33)
}
