% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateRangeFormNflFunction.R
\name{dateRangeFormNfl}
\alias{dateRangeFormNfl}
\title{labs dateRange Formatting Function}
\usage{
dateRangeFormNfl(yyyymmdd = "2014/09/03", weeks = 17)
}
\arguments{
\item{yyyymmdd}{"2018/09/2"}

\item{weeks}{any# i.e:17}
}
\value{
list of dates for use
}
\description{
labs dateRange Formatting Function
}
\examples{
dateRangeFormNfl("2018/09/05",weeks=17)
}
