% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreachRemoveColumnFunction.R
\name{foreachRemoveColumn}
\alias{foreachRemoveColumn}
\title{remove Columns Function}
\usage{
foreachRemoveColumn(d = nba1026_31, r = "|custom")
}
\arguments{
\item{d}{dataframe}

\item{r}{pattern to match in column names to be removed}
}
\value{
subbed dF
}
\description{
remove Columns Function
}
\examples{
foreach RemoveColumn(d=nba1026_31,r = "|custom")
}
