% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.definedFunction.R
\name{is.defined}
\alias{is.defined}
\title{is object defined Function}
\usage{
is.defined(sym)
}
\arguments{
\item{sym}{object to check for}
}
\value{
TRUE/FALSE
}
\description{
is object defined Function
}
\examples{
is.defined(cookies)
}
