% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labsCookiesFunction.R
\name{labsCookies}
\alias{labsCookies}
\title{lab Cookies Retrieval Functions}
\usage{
labsCookies(dockerName = "labCookies", sport = "pga")
}
\arguments{
\item{dockerName}{docker container name(can be anything)}

\item{sport}{mlb,nba,nfl,nhl,pga}
}
\value{
all FantasyLab Cookies
}
\description{
lab Cookies Retrieval Functions
}
\examples{
pgaCookies<-labsCookies("labCookies","pga")
}
