% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadedListFunction.R
\name{loadedList}
\alias{loadedList}
\title{get loaded Packages in a dataframe column sorted alphabetically for easy skim thru}
\usage{
loadedList()
}
\value{
loaded package names
}
\description{
get loaded Packages in a dataframe column sorted alphabetically for easy skim thru
}
\examples{
loadedList()
}
