% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRDataFunction.R
\name{loadRData}
\alias{loadRData}
\title{load RData saved objects function}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{mlbHitterAdvancedLm1}
}
\value{
inside of saved files
}
\description{
load RData saved objects function
}
\examples{
loadRData("~/Desktop/MLB_Daily/mlbHitterAdvancedLm1.rda")
}
