% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbArmsImportForLearningFunction.R
\name{mlbArmsImportForLearning}
\alias{mlbArmsImportForLearning}
\title{mlbArmsImportForMachineLearningFunction}
\usage{
mlbArmsImportForLearning(firstDate = "2018/04/18", number = 43,
  slateName = "Main", refreshData = FALSE)
}
\arguments{
\item{firstDate}{"yyyy/mm/dd"}

\item{number}{number of days to research}

\item{slateName}{game time to analyze over period of days researched}

\item{refreshData}{get post game data where needed..True/False.. defaults as FALSE}
}
\value{
4 dataframes to use for machine learning
}
\description{
mlbArmsImportForMachineLearningFunction
}
\examples{
mlbArmsImportForLearning("2018/04/18",number=43,slateName="Main",refreshData=FALSE)
}
