% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbArmsTrainMethodFunction.R
\name{mlbArmsTrain}
\alias{mlbArmsTrain}
\title{mlb pitcher model training function quickCaret}
\usage{
mlbArmsTrain(algo = "blackboost", dat = train.Actual.Arms)
}
\arguments{
\item{algo}{model to use}

\item{dat}{the data}
}
\value{
caret model
}
\description{
mlb pitcher model training function quickCaret
}
\examples{
mlbArmsTrain(algo='blackboost',dat=train.Actual.Arms)
}
