% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbBatsProjectionFunction2018.R
\name{mlbBatProjection}
\alias{mlbBatProjection}
\title{mlb Bat Projections Function}
\usage{
mlbBatProjection(pDate = "6_10_2018", gameSlate = "Main",
  batModel = "svmFit.actual6_11_18", labModel = "mainBae")
}
\arguments{
\item{pDate}{gameDate}

\item{gameSlate}{.."Main","Turbo","Afternoon","Allday","Night"...etc}

\item{batModel}{the model used to mak projections. Saved in ~/MLB_Daily/machineLearning/hitters/}

\item{labModel}{name of the fantasyLabs model used in the projections model training...default "mainBae"..6_11_2018}
}
\value{
saves csv file with projections to ~/MLB_Daily/dfsProjections/hitters/ and returns the data.frame in the r console
}
\description{
mlb Bat Projections Function
}
\examples{
mlbBatProjection(pDate="6_10_2018",gameSlate="Main",batModel="svmFit.actual6_11_18",labModel="mainBae")
}
