% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbCoreFunFunction.R
\name{mlbCoreFun}
\alias{mlbCoreFun}
\title{mlbCoreFun}
\usage{
mlbCoreFun(dateV = rDailyFantasy::dateRangeForm("2018/03/29", days = 1),
  partitionSize = 0.8, labModel = "2018Early", style = "projections")
}
\arguments{
\item{dateV}{vector}

\item{partitionSize}{0-1}

\item{labModel}{modelName}

\item{pastEvent}{T/F}

\item{research}{T/F}

\item{statsRefresh}{T/F}

\item{toProject}{T/F}
}
\value{
researchData or projectionData
}
\description{
mlbCoreFun
}
\examples{
mlbCoreFun()
}
