% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbModelSetupFunction.R
\name{mlbModelSetup}
\alias{mlbModelSetup}
\title{mlb Model Import Function}
\usage{
mlbModelSetup(modelDate = "3_4_2018", pastEvent = FALSE,
  labModel = "WGCBridgestone", postRefresh = NULL, parallelRun = FALSE,
  cookie = mlbCookies)
}
\arguments{
\item{modelDate}{"m_d_yyyy"}

\item{pastEvent}{TRUE/FALSE}

\item{labModel}{name of fantasyLab model to use}

\item{postRefresh}{reRun for misc postTourney stats}

\item{parallelRun}{running with doSNOW in parallel TRUE/FALSE}

\item{cookie}{labsCookies}
}
\value{
mlb Models
}
\description{
mlb Model Import Function
}
\examples{
mlbModelSetup(modelDate="4_5_2018",cookie=mlbCookies)
}
