% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naToWhatFunction.R
\name{na.What}
\alias{na.What}
\title{NA to Zero}
\usage{
na.What(x, what = 0)
}
\arguments{
\item{x}{data}

\item{what}{what to change NAs to}
}
\value{
no NA Dataframe x
}
\description{
NA to Zero
}
\examples{
na.zero(x,1)
}
