% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naToZeroFunction.R
\name{na.zero}
\alias{na.zero}
\title{NAs to 0 function}
\usage{
na.zero(x)
}
\arguments{
\item{x}{anything to replace NAs with}
}
\value{
no NA Dataframe x
}
\description{
NAs to 0 function
}
\examples{
na.zero(x)
}
