% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNbaModelFunction.R
\name{nbaModelSetup}
\alias{nbaModelSetup}
\title{NBA Player Model Retrieval Function}
\usage{
nbaModelSetup(modelDate = "10_16_2018", gameSlate = "Main",
  algoX = "xgb.actualPoints.big.full.Final",
  treatPlanX = "treatplanBig", newVarsX = "new_varsBig",
  labModel = "Money", sameDay = TRUE, systemsRefresh = 0,
  sourceRefresh = 0, cookie = nbaCookies)
}
\arguments{
\item{modelDate}{"10_16_2018"}

\item{gameSlate}{"Main"}

\item{algoX}{xgboostModel}

\item{treatPlanX}{xgboost treatplan}

\item{newVarsX}{xgboost new_vars}

\item{labModel}{"2017FirstModel}

\item{sameDay}{running on sameDay as projections date?}

\item{systemsRefresh}{0/1}

\item{sourceRefresh}{0/1}

\item{cookie}{nbaCookies}
}
\value{
labsPlayerModel
}
\description{
NBA Player Model Retrieval Function
}
\examples{
nbaModelSetup(modelDate="10_16_2018",gameSlate="Main",algoX="xgb.actualPoints.big.full.Final",treatPlanX="treatplanBig",newVarsX="new_varsBig",labModel="Money",sameDay=TRUE,systemsRefresh=0,sourceRefresh=0,cookie=nbaCookies)
}
