% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaOptimizerFunction.R
\name{nbaOptimizer}
\alias{nbaOptimizer}
\title{nbaOptimizerFunction}
\usage{
nbaOptimizer(modelDate = "11_15_2018", gameSlate = "Main",
  sameDay = TRUE, projectionSrc = "xgb", optModel = "optTrainranger",
  lineUps = 200L, bans = NULL, locks = c(""),
  nbaModel = "nbaranger2", labModel = "Money",
  zeroOwnModel = "nbaZeroOwngbmnba", labZeroModel = "Contrarians",
  cookie = nbaCookies, dCookies = dkCookies)
}
\arguments{
\item{modelDate}{date}

\item{gameSlate}{slate}

\item{sameDay}{optimizing for same day as running function? T/F}

\item{projectionSrc}{x("xgb","cubist")}

\item{optModel}{regression proj model for lineups overall}

\item{lineUps}{#lineups}

\item{bans}{exclude}

\item{locks}{include}

\item{nbaModel}{nbaranger2}

\item{labModel}{Money}

\item{zeroOwnModel}{"nbaZeroOwngbmnba"}

\item{labZeroModel}{"Contrarians"}

\item{cookie}{fantasyLabs Cookies}

\item{dCookies}{draftkings Cookies}
}
\value{
lineups
}
\description{
nbaOptimizerFunction
}
\examples{
nbaOptimizer(modelDate="11_10_2018",gameSlate="Main",sameDay=TRUE,optModel="optTrainranger",lineUps=200L,bans=NULL,locks=c(""),saveNameAdd="zeroRussellAndFlow",cookie=nbaCookies,dCookies=dkCookies)
}
