% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaOwnershipFunction.R
\name{nbaOwn}
\alias{nbaOwn}
\title{NBA ownership file}
\usage{
nbaOwn(modelDate, cookies = nbaCookies)
}
\arguments{
\item{modelDate}{}
}
\value{
ownership data
}
\description{
NBA ownership file
}
\examples{
nbaOwn(modelDate="2_8_2019",cookies=nbaCookies)
}
