% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaProjectionsFunction.R
\name{nbaProjections}
\alias{nbaProjections}
\title{nbaProjectionsFunction}
\usage{
nbaProjections(modelDate = "10_17_2018", gameSlate = "Main",
  nbaModel = "nbaCubist", zeroOwnModel = "nbaZeroOwngbmnba",
  nbaPre = "preProcNBA", labModel = "2017FirstModel",
  labZeroModel = "zeroOwnVarImp", getCorr = FALSE,
  saveNameAdd = "zeroRussellAndFlow", refreshModel = TRUE,
  cookies = nbaCookies)
}
\arguments{
\item{modelDate}{"10_17_2018"}

\item{gameSlate}{"Main"}

\item{nbaModel}{"nbaCubist"}

\item{zeroOwnModel}{zeroOwnershipProjModel}

\item{nbaPre}{"preProcNBA"}

\item{labModel}{"2017FirstModel"}

\item{labZeroModel}{fantasylabsXeroOwnModel}

\item{getCorr}{getActualCorrelelation post contests}

\item{saveNameAdd}{"russelAndFlow"}

\item{refreshModel}{T/F refreshdataSet pulled for projections?}

\item{cookies}{nbaCookies}
}
\value{
nbaProjections/also output as .csv in NBA_Daily/dfsProjections/
}
\description{
nbaProjectionsFunction
}
\examples{
nbaProjections(modelDate="10_17_2018",gameSlate="Main",nbaModel="nbaCubist",zeroOwnModel="nbaZeroOwngbmModel",nbaPre="preProcNBA",labModel="2017FirstModel",labZeroModel="zeroOwnVarImp",getCorr=FALSE,saveNameAdd="zeroRussellAndFlow",refreshModel=TRUE,cookies=nbaCookies)
}
