% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaResultsFunction.R
\name{nbaResults}
\alias{nbaResults}
\title{nbaResults Function}
\usage{
nbaResults(modelDate = "11_10_2018")
}
\arguments{
\item{modelDate}{"11_10_2018"}
}
\value{
contest Results Summaries from modelDates Contests
}
\description{
nbaResults Function
}
\examples{
nbaResults(modelDate="11_10_2018")
}
