% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaSetFactorsForProjFunction.R
\name{nbaSetFactorsForProj}
\alias{nbaSetFactorsForProj}
\title{nba Factors Setting Function for Projections}
\usage{
nbaSetFactorsForProj(train = train, target = "ActualPoints")
}
\arguments{
\item{train}{dataframe}

\item{target}{target variable for machine learning}
}
\value{
df w/ factors correctly set...
}
\description{
nba Factors Setting Function for Projections
}
\examples{
nbaSetFactorsForProj(train=train,target="ActualPoints")
}
