% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topLineupExposureScript.R
\name{nbaTop100Exposure}
\alias{nbaTop100Exposure}
\title{NBA Top 100 Lineups Player Exposure addition}
\usage{
nbaTop100Exposure(modelDate = "11_26_2018", cookies = nbaCookies)
}
\arguments{
\item{modelDate}{"11_26_2018"}

\item{cookies}{nbaCookies}
}
\value{
dataSet with 1/0 in inLineup col to indicate if player was heavily owned in top 100 winning lineups
}
\description{
NBA Top 100 Lineups Player Exposure addition
}
\examples{
nbaTop100Exposure(modelDate="11_26_2018",cookies=nbaCookies)
}
