% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflLocksFunction.R
\name{nflLocks}
\alias{nflLocks}
\title{nflLocksFunction}
\usage{
nflLocks(modelWeek = 11, modelYear = 2018, gameSlate = "Main",
  pffAdjPtsAll = TRUE)
}
\arguments{
\item{modelWeek}{11}

\item{modelYear}{2018}

\item{gameSlate}{"Main"}

\item{pffAdjPtsAll}{T/F}
}
\value{
players to lock
}
\description{
nflLocksFunction
}
\examples{
nflLocks(modelWeek=11,modelYear=2018,gameSlate="Main",pffAdjPtsAll=TRUE)
}
