% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNflModelFunction.R
\name{nflModelSetup}
\alias{nflModelSetup}
\title{Nfl Model Import Function}
\usage{
nflModelSetup(modelWeek = 1, modelYear = 2018, pastEvent = TRUE,
  labModel = "2018", postRefresh = 1, systemRefresh = 1,
  parallelRun = FALSE, qbOnly = FALSE, rbOnly = FALSE,
  wrOnly = FALSE, teOnly = FALSE, dstOnly = FALSE,
  cookie = nflCookies)
}
\arguments{
\item{modelWeek}{1:17}

\item{modelYear}{2015:2018}

\item{pastEvent}{TRUE/FALSE}

\item{labModel}{name of fantasyLab model to use}

\item{postRefresh}{reRun for misc postTourney stats(1/0)[T/F]}

\item{systemRefresh}{refresh lab models data(1/0)[T/F]}

\item{parallelRun}{running with doSNOW in parallel TRUE/FALSE}

\item{qbOnly}{T/F}

\item{rbOnly}{T/F}

\item{wrOnly}{T/F}

\item{teOnly}{T/F}

\item{dstOnly}{T/F}

\item{cookie}{labsCookies}
}
\value{
nfl Models
}
\description{
Nfl Model Import Function
}
\examples{
nflModelSetup(modelWeek=1,modelYear=2015,cookie=nflCookies)
}
