% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflPositionModelCorrelsFunction.R
\name{nflPositionModelCorrels}
\alias{nflPositionModelCorrels}
\title{nfl positionModel Correlations for Algos testing/machineLearning Models Function}
\usage{
nflPositionModelCorrels(position = "QB", modelYear = 2018,
  modelWeek = 3)
}
\arguments{
\item{position}{QB/RB/WR/TE/DST}

\item{modelYear}{2015-2018}

\item{modelWeek}{1-17}
}
\value{
list of correls for each model
}
\description{
nfl positionModel Correlations for Algos testing/machineLearning Models Function
}
\examples{
nflPositionModelCorrels(position="QB",modelYear=2018,modelWeek=3)
}
