% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflProjectionsFunction.R
\name{nflProjection}
\alias{nflProjection}
\title{nfl Player Position Projections Function}
\usage{
nflProjection(modelWeek = 6, modelYear = 2018, gameSlate = "Main",
  qbModel = "qbBig_ridgeModel_tuneUniq2",
  rbModel = "rbBig_gbmModel_tuneUniq2",
  wrModel = "wrBig_gbmModel_tuneUniq2",
  teModel = "teBig_treebagModel_tuneUniq2",
  dstModel = "dstBig_icrModel_tuneUniq2",
  pffModel = "pffCombined_rangerModel_trimmed",
  qbVars = "qbVariablesBig", rbVars = "rbVariablesBig",
  wrVars = "wrVariablesBig", teVars = "teVariablesBig",
  dstVars = "dstVariablesBig", qbPre = "qbUPre", rbPre = "rbUPre",
  wrPre = "wrUPre", tePre = "teUPre", dstPre = "dstUPre",
  pffPre = "preProcPFFDataTrimmed", labModel = "2018",
  saveNameAdd = "theMoneyMaker", pffData = TRUE)
}
\arguments{
\item{modelWeek}{mW}

\item{modelYear}{mY}

\item{gameSlate}{.."Main","Turbo","Afternoon","Allday","Night"...etc}

\item{qbModel}{the model used to mak projections. Saved in ~/MLB_Daily/machineLearning/hitters/}

\item{rbModel}{algoName}

\item{wrModel}{algoName}

\item{teModel}{algoName}

\item{dstModel}{algoName}

\item{qbVars}{variableNamesFile}

\item{rbVars}{variableNamesFile}

\item{wrVars}{variableNamesFile}

\item{teVars}{variableNamesFile}

\item{dstVars}{variableNamesFile}

\item{qbPre}{preProcessingFile}

\item{rbPre}{preProcessingFile}

\item{wrPre}{preProcessingFile}

\item{tePre}{preProcessingFile}

\item{dstPre}{preProcessingFile}

\item{labModel}{name of the fantasyLabs model used in the projections model training...default "mainBae"..6_11_2018}

\item{saveNameAdd}{addtoOriginalName}

\item{pffData}{TRUE?FALSE}
}
\value{
saves csv file with projections to ~/NFL_Daily/dfsProjections/QB/WR/TE/RB/DST/..../ and returns the data.frames in the r console in a list
}
\description{
nfl Player Position Projections Function
}
\examples{
nflProjection(modelWeek=6,modelYear=2018,gameSlate="Main",qbModel="qbBig_ridgeModel_tuneUniq2",rbModel="rbBig_gbmModel_tuneUniq2",wrModel="wrBig_gbmModel_tuneUniq2",teModel="teBig_treebagModel_tuneUniq2",dstModel="dstBig_icrModel_tuneUniq2",pffModel="pffCombined_rangerModel_trimmed",qbVars="qbVariablesBig",rbVars="rbVariablesBig",wrVars="wrVariablesBig",teVars="teVariablesBig",dstVars="dstVariablesBig",qbPre="qbUPre",rbPre="rbUPre",wrPre="wrUPre",tePre="teUPre",dstPre="dstUPre",pffPre="preProcPFFDataTrimmed",labModel="2018",saveNameAdd="theMoneyMaker",pffData=TRUE)
}
