% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflResultsFunction.R
\name{nflResults}
\alias{nflResults}
\title{nfl Contest Results Function}
\usage{
nflResults(modelYear = 2018, modelWeek = 7)
}
\arguments{
\item{modelYear}{2018}

\item{modelWeek}{1-17}
}
\value{
contest results in NFL_Daily/results directory
}
\description{
nfl Contest Results Function
}
\examples{
nflResults(2018,7)
}
