% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflTeamStatsCleanupFunction.R
\name{nflTeamStatsCleanup}
\alias{nflTeamStatsCleanup}
\title{nfl teamStats dataSet Cleanup Function}
\usage{
nflTeamStatsCleanup(modelYear = 2018, modelWeek = 4)
}
\arguments{
\item{modelYear}{2018}

\item{modelWeek}{1-17}
}
\description{
nfl teamStats dataSet Cleanup Function
}
\examples{
nflTeamStatsCleanup(modelYear=2018,modelWeek=1)
}
