% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noControlRegressionFittingFunction.R
\name{noControlModel}
\alias{noControlModel}
\title{caret model fit function}
\usage{
noControlModel(X = trainX, Y = train$ActualPoints, algo = "lm",
  pre = c("center", "scale", "YeoJohnson"), trMethod = NULL,
  seedNum = 14)
}
\arguments{
\item{X}{data}

\item{Y}{target Variable}

\item{algo}{caret method to use}

\item{pre}{preProcess params}

\item{seedNum}{seed number}

\item{para}{parallel? T/F}
}
\value{
trainModel
}
\description{
caret model fit function
}
\examples{
noControlModel(X=mlbAllBats,Y=ActualPoints,algo="lmStepAIC",seedNum=14,para=TRUE)
}
