% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericOnlyFunction.R
\name{numericOnly}
\alias{numericOnly}
\title{Numeric Only DataFrame}
\usage{
numericOnly(x)
}
\arguments{
\item{x}{}
}
\value{
data frame of numeric only type
}
\description{
Numeric Only DataFrame
}
\examples{
dfNums <- numericOnly(dF)
}
