% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzvFilterFunction.R
\name{nzv.filter}
\alias{nzv.filter}
\title{nearZeroVariance predictors function filter}
\usage{
nzv.filter(x)
}
\arguments{
\item{x}{data.frame of predictors}
}
\value{
predictors minus zeroVariance predictors from orginal set
}
\description{
nearZeroVariance predictors function filter
}
\examples{
nzv.filer(x=bats[,1:160])
}
