% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgaModelSetupFunction.R
\name{pgaModelSetup}
\alias{pgaModelSetup}
\title{pga Model Import Function}
\usage{
pgaModelSetup(modelDate = "3_4_2018", pastEvent = FALSE,
  labModel = "WGCBridgestone", postRefresh = NULL,
  parallelRun = FALSE, cookie = pgaCookies)
}
\arguments{
\item{modelDate}{"m_d_yyyy"}

\item{pastEvent}{TRUE/FALSE}

\item{labModel}{name of fantasyLab model to use}

\item{postRefresh}{reRun for misc postTourney stats}

\item{parallelRun}{running with doSNOW in parallel TRUE/FALSE}

\item{cookie}{labsCookies}
}
\value{
pga Models
}
\description{
pga Model Import Function
}
\examples{
pgaModelSetup(modelDate="4_5_2018",cookie=pgaCookies)
}
