% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectedOwnFixFunction.R
\name{projectedOwnFix}
\alias{projectedOwnFix}
\title{player Own Labs Model Fix [col]}
\usage{
projectedOwnFix(x)
}
\arguments{
\item{x}{dataframe name with p_own variable}
}
\value{
fixed ownership predictions
}
\description{
player Own Labs Model Fix [col]
}
\examples{
dataset$Properties.p_own<-projectedOwnFix(dataset)
}
