% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeColumnFunction.R
\name{removeColumn}
\alias{removeColumn}
\title{remove Column}
\usage{
removeColumn(dat = nba1026_31[[4]], remove = "|custom")
}
\arguments{
\item{dat}{dataframe}

\item{remove}{part of name of column to remove}
}
\value{
dataframe without unwanted columns
}
\description{
remove Column
}
\examples{
removeColumn(dat=nba1026_31[[4]],remove="|custom")
}
