% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCleanupFunction.R
\name{researchModelCleanup}
\alias{researchModelCleanup}
\title{ressearch Moddel Cleanup Function}
\usage{
researchModelCleanup(mod = rbAll)
}
\arguments{
\item{mod}{dataframe to evaluate}
}
\value{
preProcessed Data for mLearn
}
\description{
ressearch Moddel Cleanup Function
}
\examples{
researchModelCleanup(qbAll)
}
