% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFeatureSelectionFunction.R
\name{run.feature.selection}
\alias{run.feature.selection}
\title{run Feature Selection Function}
\usage{
run.feature.selection(num.iters = 20, feature.vars, class.var)
}
\arguments{
\item{num.iters}{iterations}

\item{feature.vars}{variables}

\item{class.var}{targetVariable}
}
\value{
rfe
}
\description{
run Feature Selection Function
}
\examples{
run.feature.selection(feature.vars=qbData$train, class.var=ActualPoints)
}
