% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPredsFunction.R
\name{testPreds}
\alias{testPreds}
\title{testPredsFunction}
\usage{
testPreds(nbaModel = nba19, wts = c(0.3, 0.6, 0.1))
}
\arguments{
\item{nbaModel}{model}

\item{wts}{c(...,...,...)}
}
\value{
stats for predictons
}
\description{
testPredsFunction
}
\examples{
testPreds(nabModel=nba19,wts=c(0.3,0.6,0.1))
}
