% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainTestSplitFunction.R
\name{trainTestSplit}
\alias{trainTestSplit}
\title{train Test Data Split for Machine Learning}
\usage{
trainTestSplit(x, sp = 0.75, targetVar = NULL)
}
\arguments{
\item{x}{dataFrame}

\item{sp}{training split percent}

\item{targetVar}{target Variable}
}
\value{
list...train,test,trainTarget,testTarget
}
\description{
train Test Data Split for Machine Learning
}
\examples{
trainTestSplit(pgaPointsAllData,sp=0.75,targetVar="ActualPoints")
}
