% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedPredictionsFunction.R
\name{weightedPredictions}
\alias{weightedPredictions}
\title{weighted Predictions}
\usage{
weightedPredictions(weights = c(0.58, 0.39), predictions = list(...))
}
\arguments{
\item{weights}{c(...,...,...)}

\item{predictions}{list(...,...,...)}
}
\value{
weighted predictions
}
\description{
weighted Predictions
}
\examples{
weightedPredctions(weights=c(0.58,0.39),predictions=list(preds,preds1))
}
