% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostTuningFunction.R
\name{xgBoostTuning}
\alias{xgBoostTuning}
\title{xgboost Tuning Script}
\usage{
xgBoostTuning(dataModel = bbmDataLearning, target = "ActualV",
  saveAsName = "bbmDataLearning")
}
\arguments{
\item{dataModel}{data}

\item{target}{regression Target Variable}

\item{saveAsName}{prefix for each saved object in NBA_Daily/machineLearning/...}
}
\value{
best tuned model
}
\description{
xgboost Tuning Script
}
\examples{
agBoostTuning(dataModel=bbmDataLearning,target="ActualV",saveAsName="bbmDataLearning")
}
