% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbModelSetupFunction.R
\name{mlbModelSetup}
\alias{mlbModelSetup}
\title{mlb Model Import Function}
\usage{
mlbModelSetup(modelDate = "4_1_2018", pastEvent = TRUE,
  labModel = "mainBae", postRefresh = 1, systemRefresh = 1,
  parallelRun = FALSE, batsOnly = FALSE, armsOnly = FALSE,
  cookie = mlbCookies)
}
\arguments{
\item{modelDate}{"m_d_yyyy"}

\item{pastEvent}{TRUE/FALSE}

\item{labModel}{name of fantasyLab model to use}

\item{postRefresh}{reRun for misc postTourney stats(1/0)[T/F]}

\item{systemRefresh}{refresh lab models data(1/0)[T/F]}

\item{parallelRun}{running with doSNOW in parallel TRUE/FALSE}

\item{batsOnly}{only hitters data}

\item{armsOnly}{only pitcher data}

\item{cookie}{labsCookies}
}
\value{
mlb Models
}
\description{
mlb Model Import Function
}
\examples{
mlbModelSetup(modelDate="4_5_2018",cookie=mlbCookies)
}
