% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaOptimizerFunction.R
\name{nbaOptimizer}
\alias{nbaOptimizer}
\title{nbaOptimizerFunction}
\usage{
nbaOptimizer(modelDate = "11_10_2018", gameSlate = "Main",
  sameDay = TRUE, lineUps = 200L, bans = NULL, locks = c(""),
  saveNameAdd = "zeroRussellAndFlow", cookie = nbaCookies,
  dCookies = dkCookies)
}
\arguments{
\item{modelDate}{date}

\item{gameSlate}{slate}

\item{sameDay}{optimizing for same day as running function? T/F}

\item{lineUps}{#lineups}

\item{bans}{exclude}

\item{locks}{include}

\item{saveNameAdd}{russellAndFlow}

\item{cookie}{fantasyLabs Cookies}

\item{dCookies}{draftkings Cookies}
}
\value{
lineups
}
\description{
nbaOptimizerFunction
}
\examples{
nbaOptimizer(modelDate="11_10_2018",gameSlate="Main",sameDay=TRUE,lineUps=200L,bans=NULL,locks=c(""),saveNameAdd="zeroRussellAndFlow",cookie=nbaCookies,dCookies=dkCookies)
}
