% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaProjectionsFunction.R
\name{nbaProjections}
\alias{nbaProjections}
\title{nbaProjectionsFunction}
\usage{
nbaProjections(modelDate = "2_8_2019", gameSlate = "Main",
  upload = "finalOptimal", algoX = "xgb.actualPoints.big.full.Final",
  treatPlanX = "treatplanBig", newVarsX = "new_varsBig",
  nbaModel = "nbaCubist", zeroOwnModel = "nbaZeroOwngbmnba",
  nbaPre = "preProcNBA", labModel = "Money",
  labZeroModel = "Contrarians", plusMinusPre = "plusMinusPreProc",
  valuePre = "valuePreProc", plusMinusModel = "plusMinusgbm",
  valueModel = "valuegbm", sameDay = TRUE, getCorr = FALSE,
  refreshModel = TRUE, cookies = nbaCookies)
}
\arguments{
\item{modelDate}{"10_17_2018"}

\item{gameSlate}{"Main"}

\item{upload}{labs Proj Upload 'finalOptimal'}

\item{algoX}{xgboost algo}

\item{treatPlanX}{xgboost treatplan}

\item{newVarsX}{xgboost new_vars}

\item{nbaModel}{"nbaCubist"}

\item{zeroOwnModel}{zeroOwnershipProjModel}

\item{nbaPre}{"preProcNBA"}

\item{labModel}{"2017FirstModel"}

\item{labZeroModel}{fantasylabsXeroOwnModel}

\item{plusMinusPre}{"plusMinusPreProc"}

\item{valuePre}{"valuePreProc"}

\item{plusMinusModel}{"plusMinusgbm"}

\item{valueModel}{"valuegbm"}

\item{sameDay}{T/F}

\item{getCorr}{getActualCorrelelation post contests}

\item{refreshModel}{T/F refreshdataSet pulled for projections?}

\item{cookies}{nbaCookies}
}
\value{
nbaProjections/also output as .csv in NBA_Daily/dfsProjections/
}
\description{
nbaProjectionsFunction
}
\examples{
nbaProjections(modelDate="2_8_2019",gameSlate="Main",upload='finalOptimal',algoX="xgb.actualPoints.big.full.Final",treatPlanX="treatplanBig",newVarsX="new_varsBig",nbaModel="nbaCubist",zeroOwnModel="nbaZeroOwngbmnba",nbaPre="preProcNBA",labModel="Money",labZeroModel="Contrarians",plusMinusPre="plusMinusPreProc",valuePre="valuePreProc",plusMinusModel="plusMinusgbm",valueModel="valuegbm",sameDay=TRUE,getCorr=FALSE,refreshModel=TRUE,cookies=nbaCookies)
}
